KAFKA-9183; Remove redundant admin client integration testing
